Act074_Next_Tickets_Adapter.java Act076_Main.java     - Variaveis de tradução @#$ Act075_Product_List_Adapter.java     - Variaveis de tradução.     - Acerto na desabilitação de UI. @#$ TK_Ticket_Ctrl.java WS_TK_Ticket_Download.java     - Acerto do Daniel, feito para continuar com o trabalho e evitar conflito no merge @#$ Act_Product_Selection_Presenter.java Sql_Act007_004.java     - Alteração para lista apenas produtos e não grupos. Validar com a chefia. @#$ Sql_Act007_003.java     - query inoeprante, vou salvar pois corrige erros anteriores. @#$ Act068_Main.java     - Sincroniza ticket antes de buscar pelos proximos tickets. @#$ Act075_Main.java Act075_Main_Contract.java Act075_Main_Presenter.java     - Tratamento de saída da tela de produtos quando há alteração não salva     - Save offline do produto. (Verificar como será feito). @#$ namoa_cell_12_states.xml     - Botão save da tela de produtos. @#$ act070_pipeline_header.xml     - Remoção de icone de atalho de origem     - REgra de allcaps para botão do sincronismo @#$ act075_main_content.xml     - comportamento novo para botão de save     - novo tratamento para wrap constraint. @#$

+++ app/src/main/java/com/namoadigital/prj001/service/WS_TK_Ticket_Download.java
@@ -128,7 +128,10 @@ public class WS_TK_Ticket_Download extends IntentService {
             HMAux hmAux = new HMAux();
             for (TK_Ticket tkTicket : ticketList) {
                 tkTicket.setPK();
-                daoObjReturn = ticketDao.removeFull(tkTicket);
+                TK_Ticket dbTicket = getDbTicket(tkTicket);
+                if(dbTicket != null) {
+                    daoObjReturn = ticketDao.removeFull(tkTicket);
+                }
                 //
                 if(daoObjReturn.hasError()) {
                     break;
